backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Resume Parsing (minimal)
pdfplumber==0.10.3
python-docx==1.1.0

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
